+++ src/main/java/com/javarush/island/zhidebaev/animals/Wolf.java
@@ -6,7 +6,7 @@ import com.javarush.island.zhidebaev.entities.Predators;
 
 public class Wolf extends Animal implements Predators {
     private final String icon= "\uD83D\uDC3A";
-    public Wolf(int weight, int maxNumberOfSpecies, int speed, int satiety){
+    public Wolf(double weight, double maxNumberOfSpecies, double speed, double satiety){
         super(weight, maxNumberOfSpecies, speed, satiety);
     }
